+++ delphin/semi.py
@@ -15,7 +15,7 @@ representations.
 
 import re
 from os.path import dirname, join as pjoin
-from collections import namedtuple
+from operator import itemgetter
 import warnings
 
 from delphin import tfs
@@ -66,7 +66,7 @@ _predicate_entry_re = re.compile(
 
 _synopsis_re = re.compile(
     r'\s*(?P<optional>\[\s*)?'
-    r'(?P<role>[^ ]+) (?P<value>[^ ,.{\]]+)'
+    r'(?P<name>[^ ]+) (?P<value>[^ ,.{\]]+)'
     r'(?:\s*\{\s*(?P<properties>[^ ]+ [^ ,}]+(?:, [^ ]+ [^ ,}]+)*)\s*\})?'
     r'(?(optional)\s*\])'
     r'(?:\s*(?:,\s*|$))',
@@ -86,25 +86,6 @@ class SemIWarning(PyDelphinWarning):
     """
 
 
-SynopsisRole = namedtuple('SynopsisRole', 'name value properties optional')
-SynopsisRole.__doc__ = """
-    Role data associated with a SEM-I predicate synopsis.
-
-    Args:
-        name (str): the role name
-        value (str): the role value (variable type or "string")
-        properties (list): default properties associated with the
-            role's value as a list of (propname, propval) pairs
-        optional (bool): a flag indicating if the role is optional
-    Attributes:
-        name: the role name
-        value: the role value (variable type or "string")
-        properties: default properties associated with the role's
-            value as a list of (propname, propval) pairs
-        optional: a flag indicating if the role is optional
-    """
-
-
 def load(fn, encoding='utf-8'):
     """
     Read the SEM-I beginning at the filename *fn* and return the SemI.
@@ -116,6 +97,7 @@ def load(fn, encoding='utf-8'):
     Returns:
         The SemI defined by *fn*
     """
+
     data = _read_file(fn, dirname(fn), encoding)
     return SemI(**data)
 
@@ -226,9 +208,108 @@ def _read_file(fn, basedir, encoding):
                             if pair.strip() != '')
                         d['optional'] = bool(d['optional'])
                         roles.append(d)
-                    data['predicates'][pred]['synopses'].append(roles)
+                    data['predicates'][pred]['synopses'].append(
+                        {'roles': roles})
 
     return data
+
+
+class SynopsisRole(tuple):
+    """
+    Role data associated with a SEM-I predicate synopsis.
+
+    Args:
+        name (str): the role name
+        value (str): the role value (variable type or `"string"`)
+        properties (dict): properties associated with the role's value
+        optional (bool): a flag indicating if the role is optional
+    Example:
+
+    >>> role = SynopsisRole('ARG0', 'x', {'PERS': '3'}, False)
+    """
+
+    def __new__(cls, name, value, properties=None, optional=False):
+        if not properties:
+            properties = {}
+        else:
+            properties = {prop.upper(): val.lower()
+                          for prop, val in dict(properties).items()}
+        return super().__new__(cls, ([name.upper(),
+                                      value.lower(),
+                                      properties,
+                                      bool(optional)]))
+
+    name = property(itemgetter(0), doc='The role name.')
+    value = property(
+        itemgetter(1), doc='The role value (variable type or "string"')
+    properties = property(itemgetter(2), doc='Property-value map.')
+    optional = property(itemgetter(3), doc="`True` if the role is optional.")
+
+    def _to_dict(self):
+        d = {"name": self.name, "value": self.value}
+        if self.properties:
+            d['properties'] = dict(self.properties)
+        if self.optional:
+            d['optional'] = True
+        return d
+
+    @classmethod
+    def _from_dict(cls, d):
+        return cls(d['name'],
+                   d['value'],
+                   d.get('properties', []),
+                   d.get('optional', False))
+
+
+class Synopsis(tuple):
+    """
+    A SEM-I predicate synopsis.
+
+    A synopsis describes the roles of a predicate in a semantic
+    structure, so it is no more than a tuple of roles as
+    :class:`SynopsisRole` objects. The length of the synopsis is thus
+    the arity of a predicate while the individual role items detail
+    the role names, argument types, associated properties, and
+    optionality.
+    """
+
+    @classmethod
+    def from_dict(cls, d):
+        """
+        Create a Synopsis from its dictionary representation.
+
+        Example:
+
+        >>> synopsis = Synopsis.from_dict({
+        ...     'roles': [
+        ...         {'name': 'ARG0', 'value': 'e'},
+        ...         {'name': 'ARG1', 'value': 'x',
+        ...          'properties': {'NUM': 'sg'}}
+        ...     ]
+        ... })
+        ...
+        >>> len(synopsis)
+        2
+        """
+        return cls(SynopsisRole._from_dict(role)
+                   for role in d.get('roles', []))
+
+    def to_dict(self):
+        """
+        Return a dictionary representation of the Synopsis.
+
+        Example:
+
+        >>> Synopsis([
+        ...     SynopsisRole('ARG0', 'e'),
+        ...     SynopsisRole('ARG1', 'x', {'NUM': 'sg'})
+        ... ]).to_dict()
+        {'roles': [{'name': 'ARG0', 'value': 'e'},
+                   {'name': 'ARG1', 'value': 'x',
+                    'properties': {'NUM': 'sg'}}]}
+        """
+
+        return {'roles': [role._to_dict() for role in self]}
 
 
 class SemI(object):
@@ -256,6 +337,7 @@ class SemI(object):
             :class:`~delphin.tfs.TypeHierarchyNode` contains a list of
             synopses
     """
+
     def __init__(self,
                  variables=None,
                  properties=None,
@@ -308,78 +390,80 @@ class SemI(object):
             pred = pred.lower()
             parents = data.get('parents')
             synopses = []
-            for synopsis_data in data['synopses']:
+            for synopsis_data in data.get('synopses', []):
                 synopses.append(
                     self._init_synopsis(pred, synopsis_data, propcache))
             _add_to_subhierarchy(subhier, pred, parents, synopses)
         self.predicates.update(subhier)
 
     def _init_synopsis(self, pred, synopsis_data, propcache):
-        synopsis = []
-        for d in synopsis_data:
-            role = d['role'].upper()
-            value = d['value'].lower()
-            if role not in self.roles:
-                raise SemIError('{}: undefined role: {}'.format(pred, role))
-            if value == STRING_TYPE:
-                if d.get('properties', False):
+        synopsis = Synopsis.from_dict(synopsis_data)
+        for role in synopsis:
+            if role.name not in self.roles:
+                raise SemIError(
+                    '{}: undefined role: {}'.format(pred, role.name))
+            if role.value == STRING_TYPE:
+                if role.properties:
                     raise SemIError('{}: strings cannot define properties'
                                     .format(pred))
-            elif value not in self.variables:
+            elif role.value not in self.variables:
                 raise SemIError('{}: undefined variable type: {}'
-                                .format(pred, value))
+                                .format(pred, role.value))
             else:
-                props = {}
-                for k, v in dict(d.get('properties', [])).items():
-                    k, v = k.upper(), v.lower()
+                for k, v in role.properties.items():
                     if v not in self.properties:
                         raise SemIError(
                             '{}: undefined property value: {}'
                             .format(pred, v))
-                    if k not in propcache[value]:
+                    if k not in propcache[role.value]:
                         # Just warn because of the current situation where
                         # 'i' variables are used for unexpressed 'x's
                         warnings.warn(
                             "{}: property '{}' not allowed on '{}'"
-                            .format(pred, k, value),
+                            .format(pred, k, role.value),
                             SemIWarning)
                     else:
-                        _v = propcache[value][k]
+                        _v = propcache[role.value][k]
                         if not self.properties.compatible(v, _v):
                             raise SemIError(
                                 '{}: incompatible property values: {}, {}'
                                 .format(pred, v, _v))
-                    props[k] = v
-            synopsis.append(
-                SynopsisRole(role, value, props, d.get('optional', False)))
         return synopsis
 
     @classmethod
     def from_dict(cls, d):
         """Instantiate a SemI from a dictionary representation."""
+
         return cls(**d)
 
     def to_dict(self):
         """Return a dictionary representation of the SemI."""
-        def parents(node):
+
+        def add_parents(d, node):
             ps = node.parents
-            if ps == [TOP_TYPE]:
-                return []
-            return ps
+            if ps and ps != [TOP_TYPE]:
+                d['parents'] = ps
 
-        variables = {var: {'parents': parents(node),
-                           'properties': [[k, v] for k, v in node.data]}
-                     for var, node in self.variables.items()}
-        properties = {prop: {'parents': parents(node)}
-                      for prop, node in self.properties.items()}
+        variables = {}
+        for var, node in self.variables.items():
+            variables[var] = d = {}
+            add_parents(d, node)
+            if node.data:
+                d['properties'] = [[k, v] for k, v in node.data]
+
+        properties = {}
+        for prop, node in self.properties.items():
+            properties[prop] = d = {}
+            add_parents(d, node)
+
         roles = {role: {'value': value} for role, value in self.roles.items()}
-        predicates = {}
-        _synopsis_fields = 'role value properties optional'.split()
 
+        predicates = {}
         for pred, node in self.predicates.items():
-            synopses = [[dict(zip(_synopsis_fields, d)) for d in synopsis]
-                        for synopsis in node.data]
-            predicates[pred] = {'parents': parents(node), 'synopses': synopses}
+            predicates[pred] = d = {}
+            add_parents(d, node)
+            if node.data:
+                d['synopses'] = [synopsis.to_dict() for synopsis in node.data]
 
         return {'variables': variables,
                 'properties': properties,
@@ -413,6 +497,7 @@ class SemI(object):
             [('ARG0', 'e', [], False), ('ARG1', 'i', [], False),
              ('ARG2', 'i', [], False)]
         """
+
         if predicate not in self.predicates:
             raise SemIError('undefined predicate: {}'.format(predicate))
         if roles is not None: